Move Supabase Remote Storage to Connectors module
